requirements.txt
numpy 
pandas
setuptools
google-cloud-storage
pyyaml
tensorflow
joblib
comet_ml
dvc